+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: DERYCKE <DERYCKE@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/25 03:09:46 by DERYCKE           #+#    #+#             */
-/*   Updated: 2018/05/30 15:27:49 by DERYCKE          ###   ########.fr       */
+/*   Updated: 2018/06/06 21:18:10 by DERYCKE          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ int					ft_isascii(int c);
 
 //Affichage char + conversion
 int					ft_atoi(const char *str);
-char				*ft_itoa(int n);
+char				*ft_itoa(long int n);
 void				ft_putchar(char c);
 void				ft_putchar_fd(char c, int fd);
 void				ft_putnbr(int n);
